Enable checking both block and inodes if neither is specified.
This changes the behaviour around executing `check_disk` without a `-i`
or `-b` parameter.

Previously you would recieve a message telling you to pass one of those
paramters. Now, it will scan through blocks, then inodes. While
scanning through blocks or inodes it will exit with a message if usage is above
warning or critical thresholds. If neither blocks or inodes exits we
will exit with 'ok' and a message saying we are below the warning
parameter.

Executing with '-i' or '-b' and will result in the same behaviour as
before this change.