exit critical first.
A disk that is critical, will also be a disk that is warning.

A disk that is warning, is not a disk that is critical.

We could guard warning with something more than `if disk.warning?`; maybe tack a `&& ! disk.critical?`, but flipping
the order of these should also achieve the same without extra complexity.